# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Senne Platform Documentation'
  version: 1.0.0
paths:
  '/api/login/{role_id}':
    post:
      tags:
        - Login
      summary: 'User Login'
      description: 'Login User Here'
      operationId: authLogin
      parameters:
        -
          name: role_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: email
                password:
                  type: password
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'Login Successfully'
          content:
            application/json:
              schema: []
        200:
          description: 'Login Successfully'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        404:
          description: 'Resource Not Found'
  '/api/auth/login/{id}':
    post:
      tags:
        - RegisterUserMaster
      summary: 'Register User Master'
      description: 'Create the User Master Here'
      operationId: authRegister
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                name:
                  type: text
                email:
                  type: email
                cpf:
                  type: text
                cnpj:
                  type: text
                phone:
                  type: text
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'User Master Created Successfully'
          content:
            application/json:
              schema: []
        200:
          description: 'User Master Created Successfully'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
  /api/auth/logout:
    post:
      tags:
        - Logout
      summary: Logout
      description: Logout
      operationId: Logout
      responses:
        201:
          description: 'Logged out Successfully'
          content:
            application/json:
              schema: []
        200:
          description: 'Logged out  Successfully'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        404:
          description: 'Resource Not Found'
  /api/password/forgot:
    post:
      tags:
        - ResetPassword
      summary: 'Reset Password'
      description: 'Reset Password'
      operationId: ResetPassword
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                email:
                  type: email
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'Mail sended'
          content:
            application/json:
              schema: []
        200:
          description: 'Mail sended'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
...